Total intergenic pi calculation script
A script which takes as input a table of SNPs from intergenic regions of the virus genome of interest, and a gtf file of the CDS. It uses these to calculate, first, per site nucleotide diversity tables for each virus, and then intergenic pi for the virus. 
This script is called by the diversity calculations script and should be in the same folder.